# **Track Record 6 <u>Modify Optize2</u>**： Commits on Apr 09, 2026

+++ NexUs/Source/DeveloperComponents/NXBreadcrumbBarDelegate.cpp
@@ -8,14 +8,18 @@ NXBreadcrumbBarDelegate::NXBreadcrumbBarDelegate(QObject *parent)
     : QStyledItemDelegate { parent }
 {
   _themeMode = nxTheme->getThemeMode();
-  connect(nxTheme, &NXTheme::themeModeChanged, this, [=](NXThemeType::ThemeMode themeMode) { _themeMode = themeMode; });
+  connect(nxTheme, &NXTheme::themeModeChanged, this, [=](NXThemeType::ThemeMode themeMode)
+  {
+    _themeMode = themeMode;
+  });
 }
 
-NXBreadcrumbBarDelegate::~NXBreadcrumbBarDelegate() { }
+NXBreadcrumbBarDelegate::~NXBreadcrumbBarDelegate()
+{
+}
 
-void NXBreadcrumbBarDelegate::paint(QPainter *painter,
-                                    const QStyleOptionViewItem& option,
-                                    const QModelIndex& index) const
+void
+NXBreadcrumbBarDelegate::paint(QPainter *painter, const QStyleOptionViewItem &option, const QModelIndex &index) const
 {
   painter->save();
   painter->setRenderHints(QPainter::Antialiasing | QPainter::TextAntialiasing);
@@ -33,7 +37,10 @@ void NXBreadcrumbBarDelegate::paint(QPainter *painter,
     else
     {
       // 不为最后一个 且没有被鼠标覆盖
-      if (!(option.state & QStyle::State_MouseOver)) { painter->setPen(NXThemeColor(_themeMode, BasicTextNoFocus)); }
+      if (!(option.state & QStyle::State_MouseOver))
+      {
+        painter->setPen(NXThemeColor(_themeMode, BasicTextNoFocus));
+      }
     }
   }
   if (breadcrumbDisplayData != QStringLiteral(">"))
@@ -47,7 +54,7 @@ void NXBreadcrumbBarDelegate::paint(QPainter *painter,
     iconFont.setPixelSize(painter->font().pixelSize() * 0.785);
     painter->setFont(iconFont);
     itemRect.setX(itemRect.x() - itemRect.width() * 0.36);
-    painter->drawText(itemRect, Qt::AlignCenter, QChar((unsigned short) NXIconType::AngleRight));
+    painter->drawText(itemRect, Qt::AlignCenter, QChar(NXIconType::AngleRight));
   }
   painter->restore();
 }
